+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/annotation/SQLDatabase.java
@@ -113,4 +113,26 @@ public @interface SQLDatabase {
         String[] additionalTables();
         boolean reserveOriginalColumns() default false;
     }
+
+    /**
+     * Support add an extra column from the parent classes as a column of the additional table.
+     * This column would be created in both the primary and additional tables.
+     * Notice: This annotation should be declared on the leaf subclasses.
+     */
+    @Target({ElementType.TYPE})
+    @Retention(RetentionPolicy.RUNTIME)
+    @Repeatable(MultipleExtraColumn4AdditionalEntity.class)
+    @interface ExtraColumn4AdditionalEntity {
+        String additionalTable();
+        String parentColumn();
+    }
+
+    /**
+     * The support of the multiple {@link ExtraColumn4AdditionalEntity}s on the class.
+     */
+    @Target({ElementType.TYPE})
+    @Retention(RetentionPolicy.RUNTIME)
+    @interface MultipleExtraColumn4AdditionalEntity {
+        ExtraColumn4AdditionalEntity[] value();
+    }
 }