+++ docs/en/changes/changes.md
@@ -59,6 +59,9 @@
 * [Breaking Change] rename configuration folder from `otel-oc-rules` to `otel-rules`.
 * [Breaking Change] rename configuration field from `enabledOcRules` to `enabledOtelRules` and
   environment variable name from `SW_OTEL_RECEIVER_ENABLED_OC_RULES` to `SW_OTEL_RECEIVER_ENABLED_OTEL_RULES`.
+* [Breaking Change] Fix JDBC TTL to delete additional tables data. 
+  SQL Database requires removing `segment`,`segment_tag`, `logs`, `logs_tag`, `alarms`, `alarms_tag`, `zipkin_span`, `zipkin_query` before OAP starts.
+* SQL Database: add `@SQLDatabase.ExtraColumn4AdditionalEntity` to support add an extra column from parent to an additional table.
 
 #### UI